server/requirements.txt
flask==3.0.0
pyyaml==6.0.1
docker==7.0.0
requests==2.31.0
gunicorn==21.2.0
